refactored SignTransactionsModals
